Moved MonetaryAmount
