+++ Projects/Forever/Sources/Service/DemoImplementation/ForeverClientDemo.swift
@@ -1,3 +1,5 @@
+import hCore
+
 public class ForeverClientDemo: ForeverClient {
     var code: String = "CODE"
     public init() {}
